Update app.js
In the provided JavaScript file, the issue seems to be with the Google Chart API not supporting large dimensions for QR codes. When the dimension value exceeds a certain limit, the QR code does not generate as expected. To resolve this issue, we can switch to a different QR code generation library that supports larger dimensions. One of the popular libraries for generating QR codes is qrcode.js. Below is the updated JavaScript file (app.js) using the qrcode.js library

#14