Final loader & QR error solved app.js
In the provided JavaScript file, the issue seems to be with the Google Chart API not supporting large dimensions for QR codes. When the dimension value exceeds a certain limit, the QR code does not generate as expected. To resolve this issue, we can switch to a different QR code generation library that supports larger dimensions. One of the popular libraries for generating QR codes is qrcode.js. Below is the updated JavaScript file (app.js) using the qrcode.js library

Based on the provided files, an additional improvement can be made to enhance the user experience by adding a loading indicator while the QR code is being generated. To achieve this, you can add a simple loading spinner to indicate that the QR code is being processed. Here is the updated code to include the loading indicator:

#14